// frontend/src/app/codinglab-api/enums/userErrorCodes.ts
export enum UserErrorCode {
  InvalidLogin = 'InvalidLogin',
  EmailInUse = 'EmailInUse',
  ConfirmYourEmail = 'ConfirmYourEmail',
  InvalidEmail = 'InvalidEmail',
  InvalidFirstName = 'InvalidFirstName',
  InvalidLastName = 'InvalidLastName',
  InvalidRole = 'InvalidRole',
  InvalidLanguage = 'InvalidLanguage',
  InvalidPassword = 'InvalidPassword',
  RecentlySentVerificationEmail = 'RecentlySentVerificationEmail',
}