// frontend/src/app/codinglab-api/enums/jobErrorCodes.ts
export enum JobErrorCodes {
  // apply job
  InvalidJobId = 'InvalidJobId',
  InvalidStatus = 'InvalidStatus',
  InvalidHourlyRate = 'InvalidHourlyRate',
  InvalidResume = 'InvalidResume',
  InvalidMotivation = 'InvalidMotivation',
  InvalidPhone = 'InvalidPhone',
  AlreadyApplied = 'AlreadyApplied',
  JobNotAuthorized = 'JobNotAuthorized',
}